+++ app/src/main/java/org/simple/clinic/activity/placeholder/PlaceholderScreen.kt
@@ -1,50 +1,54 @@
 package org.simple.clinic.activity.placeholder
 
-import android.content.Context
-import android.util.AttributeSet
-import android.widget.RelativeLayout
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.View.VISIBLE
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
 import kotlinx.parcelize.Parcelize
-import org.simple.clinic.R
 import org.simple.clinic.await.Await
 import org.simple.clinic.await.Checkpoint
 import org.simple.clinic.databinding.ScreenPlaceholderBinding
-import org.simple.clinic.router.screen.FullScreenKey
+import org.simple.clinic.navigation.v2.ScreenKey
 import java.util.concurrent.TimeUnit.SECONDS
 
-class PlaceholderScreen(context: Context, attrs: AttributeSet) : RelativeLayout(context, attrs) {
+class PlaceholderScreen : Fragment() {
   private val delayToShowMessage = SECONDS.toMillis(3).toInt()
   private val await = Await(listOf(Checkpoint.unit(delayToShowMessage)))
   private var awaitDisposable: Disposable? = null
-  private var binding: ScreenPlaceholderBinding? = null
+
+  private var _binding: ScreenPlaceholderBinding? = null
+  private val binding get() = _binding!!
 
   private val loadingTextLayout
-    get() = binding!!.loadingTextLayout
+    get() = binding.loadingTextLayout
 
   private val loadingProgressBar
-    get() = binding!!.loadingProgressBar
+    get() = binding.loadingProgressBar
 
-  override fun onFinishInflate() {
-    super.onFinishInflate()
-    if (isInEditMode) {
-      return
-    }
-
-    binding = ScreenPlaceholderBinding.bind(this)
+  override fun onCreateView(
+      inflater: LayoutInflater,
+      container: ViewGroup?,
+      savedInstanceState: Bundle?
+  ): View {
+    _binding = ScreenPlaceholderBinding.inflate(layoutInflater, container, false)
+    return binding.root
   }
 
-  override fun onAttachedToWindow() {
-    super.onAttachedToWindow()
+  override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+    super.onViewCreated(view, savedInstanceState)
     awaitDisposable = await.items()
         .observeOn(AndroidSchedulers.mainThread())
         .subscribe { showLoadingUi() }
   }
 
-  override fun onDetachedFromWindow() {
+  override fun onDestroyView() {
+    super.onDestroyView()
+    _binding = null
     awaitDisposable?.dispose()
-    binding = null
-    super.onDetachedFromWindow()
   }
 
   private fun showLoadingUi() {
@@ -53,10 +57,10 @@ class PlaceholderScreen(context: Context, attrs: AttributeSet) : RelativeLayout(
   }
 
   @Parcelize
-  object PlaceHolderScreenKey : FullScreenKey {
+  data class Key(
+      override val analyticsName: String = "Placeholder Screen"
+  ) : ScreenKey() {
 
-    override val analyticsName = "Placeholder Screen"
-
-    override fun layoutRes() = R.layout.screen_placeholder
+    override fun instantiateFragment() = PlaceholderScreen()
   }
 }